Remove interceptors to seperat module
