+++ components/micro-gateway-cli/src/main/resources/templates/service.mustache
@@ -39,7 +39,7 @@ import wso2/gateway;
     {{#if apiRequestInterceptor}}
     int {{operationId}}_api_request_interceptor_index = -1;
     {{/if}}
-    {{#if apiRequestInterceptor}}
+    {{#if apiResponseInterceptor}}
     int {{operationId}}_api_response_interceptor_index = -1;
     {{/if}}
 
@@ -307,7 +307,7 @@ function initInterceptorIndexes{{cut qualifiedServiceName " "}}() {
             {{operationId}}_api_request_interceptor_index = gateway:loadInterceptorClass({{operationId}}_api_request_interceptor_name.substring(5, {{operationId}}_api_request_interceptor_name.length()));
         }
         {{/if}}
-        {{#if apiRequestInterceptor}}
+        {{#if apiResponseInterceptor}}
         string {{operationId}}_api_response_interceptor_name = "{{apiResponseInterceptor}}";
         if({{operationId}}_api_response_interceptor_name.startsWith("java:")) {
             {{operationId}}_api_response_interceptor_index = gateway:loadInterceptorClass({{operationId}}_api_response_interceptor_name.substring(5, {{operationId}}_api_response_interceptor_name.length()));
